requirements.txt
# Core MLOps
zenml>=0.70.0
mlflow>=2.10.0
click>=8.1.0

# Deep Learning (CPU, stable, Python 3.11 compatible)
tensorflow>=2.15.0,<2.16.0
numpy>=1.24.0,<2.0.0

# Data Processing
pandas>=2.1.0
scikit-learn>=1.3.0

# Validation & Utilities
deepchecks[vision]>=0.18.0
s3fs>=2023.12.0
Pillow>=10.0.0
pyyaml>=6.0.1
